[#2] Base logic implementation
- added layouts
- done small project's changes
