+++ src/layouts/AuthLayout.vue
@@ -1,0 +1,13 @@
+<template>
+  <router-view/>
+</template>
+
+<script>
+export default {
+name: "AuthLayout"
+}
+</script>
+
+<style scoped>
+
+</style>
