+++ src/layouts/MainLayout.vue
@@ -1,0 +1,13 @@
+<template>
+  <router-view/>
+</template>
+
+<script>
+export default {
+  name: "MainLayout"
+}
+</script>
+
+<style scoped>
+
+</style>
