+++ src/router.js
@@ -1,0 +1,34 @@
+import Vue from 'vue'
+import VueRouter from 'vue-router'
+import Home from './views/Home.vue'
+
+Vue.use(VueRouter)
+
+const routes = [
+  {
+    path: '/',
+    name: 'Home',
+    meta: {layout: 'main'},
+    component: Home
+  },
+  {
+    path: '/sign_in',
+    name: 'SignIn',
+    meta: {layout: 'auth'},
+    component: () => import('./views/SignIn.vue')
+  },
+  {
+    path: '/podcasts',
+    name: 'PodcastList',
+    meta: {layout: 'main'},
+    component: () => import('./views/PodcastList.vue')
+  }
+]
+
+const router = new VueRouter({
+  mode: 'history',
+  base: process.env.BASE_URL,
+  routes
+})
+
+export default router
